add HTTP methods to controllers
