+++ src/Controller/StyleGuideController.php
@@ -9,7 +9,7 @@ use Symfony\Component\Routing\Annotation\Route;
 class StyleGuideController extends AbstractController
 {
   // STYLE GUIDE
-  #[Route('/styleguide', name: 'style_guide')]
+  #[Route('/styleguide', name: 'style_guide', methods: ['GET'])]
   public function styleGuide(): Response
   {
     return $this->render('style_guide/styleGuide.html.twig');
