+++ templates/trick/displayOne.html.twig
@@ -135,7 +135,7 @@
     <div data-close-modal="" class="close-modal"><i class="fas fa-times"></i></div>
     <strong>Voulez-vous vraiment supprimer le trick ?</strong>
     <div class="modal-btn-container">
-      <form id="modal_form" method="POST">
+      <form id="modal_form">
         <button id="modal_submit_btn">Oui</button>
         <button data-close-modal="">Non</button>
       </form>
